TST/CLN: Split out some to_string tests
